Move enable_local_pickup_without_address so it works under all contexts
